First Website Version
[Problem]
Nothing in the project!

[Solution]
Added the baseline project. Contains a simple website that can be
browsed through.

[Testing]
Manual testing currently. Will investigate automated testing at a later
time. Note that when testing on a local web browser, Firefox must be
used! This is due to how the config.json file is loaded!