Timeline Page Added
[Problem]
Timeline page doesn't exist! On top of this, there's a few bugs on
Microsoft Edge and Internet Explorer.

[Solution]
Fixed a bug in Microsoft Edge and Internet Explorer where the footer had
a large gap between it and the bottom of the page. Also fixed the social
media icons in Internet Explorer to properly behave when you hover over
them.

Added a new timeline page that can be updated from a list of events in
the config.json file. Can be set to order chronologically or backwards,
allowing for users to best present their job history (or education
history).

[Testing]
Tested across Google Chrome, Mozilla Firefox, Microsoft Edge, and
Internet Explorer, as well as on my mobile Android phone using Google
Chrome.