Projects Page + Bug Fix
[Problem]
Projects page is empty and the timeline is cut off on Firefox for
Android.

[Solution]
Fixed the Firefox for Android bug by adding some additional padding to
the bottom of the page body.

Added a new section to the projects page. Loads projects from Github
(see username in Config file) and creates cards out of them that users
can then read through. Includes links to the repo as well as download
links of releases are available. Also added a way to load custom
projects, defined in a similar format. Repos from both custom projects
and GitHub can be combined, and will be sorted by the last upated date.

[Tetsing]
Tested across several browsers, ensured that everything looks fine and
works fine.