Speedups and Cleanups
[Problem]
The projects page can be a bit slow to show anything to users sync
everything is loaded synchronously currently. On top of this, there are
no default images for when there are no timeline photos provided, and
the site is currently configured all in one configuration file.

[Solution]
First, setup an asynchronous loading of all the projects. This should
quickly add some of the projects to the page, lower
user-percieved-latency.

Second, added a bunch of shiny new icons from GitHub's Octicons for
default images, and ensured they're configurable so that users can
customize them.

Finally, split out the configuration files into the configs folder. This
will make it easier to add bits and pieces to a new setup. Also updated
the README to be a bit more concise about all of this.

[Testing]
Manually tested across multiple web browsers. Someday I'll add automated
testing to this.