Combined Media + Projects Fixed
[Problem]
One of my projects had an incorrect video link. On top of this,
currently images are a separate button, and only support a single image.

[Solution]
Changed the video modal to be a media modal, and added support for
images. Removed the old image button for simplification. Updated Rift
Poke's video URL (used MP3 previously for testing). Media for projects
is now complete!

[Testing]
Tested in Firefox and Chrome. Confident this shouldn't break anything.