RSS Image Links
[Problem]
When looking at the RSS feed cards on the home page, there's an instinct
to try and click the images to read the articles. These should probably
be links instead to the article.

[Solution]
Setup the images to be links to the articles instead of just images.

[Testing]
Quick test in a web browser. Minor changes so there shouldn't be any
compatibility issues.