SEO Improvements
[Problem]
There are some issues with how SEO (Search Engine Optimizations) pick up
the sites content. Keywords alone are not enough.

[Solution]
First, updated the keywords for my own. Second, set the title of the
home page to just be the user's name instead. Third, changed the home
page banner text to "Welcome to X's website!" instead of just
"Welcome!". This gets the user's name front and center. Finally, added a
"bio_brief" configuration value for a brief biography to display on the
home page. This ensures that information about the user is picked up by
search engines.

[Testing]
Quick set of testing to ensure everything appears in a web browser.