Add Credits Block + Cleanups
[Problem]
No credits section on the Contact Page. On top of this, some minor bugs
with how the Experience Page header is laid out, plus the sitemap
is out of date.

[Solution]
Added a small credits section to the Contact Page, updated the Experience
Page to have a properly centered header, plus updated the Projects Page
to use a top margin instead of a break for its header. Finally, updated
the sitemap to be accurate with the lastmod dates.

[Testing]
Quick set of testing in Chrome. Changes should not affect other browsers.