New Project + Image Size Cap
[Problem]
New project to add plus the images on the projects page should have a
max height in pixels, that way they are never too large for the page.

[Solution]
Set the max height for images to be 500px for the time being (may
increase it or decrease it later depending on feedback), plus added the
new project. Updated the .gitignore to ignore video files too.

[Testing]
Quick test in FireFox.